requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
PyPDF2
torch